+++ src/renderer/utils/musicSdk/kg/pic.js
@@ -37,13 +37,12 @@ export default {
         },
       },
     )
-    requestObj.promise = requestObj.promise.then(({ body }) => {
+    return requestObj.promise.then(({ body }) => {
       if (body.error_code !== 0) return Promise.reject('图片获取失败')
       let info = body.data[0].info
       const img = info.imgsize ? info.image.replace('{size}', info.imgsize[0]) : info.image
       if (!img) return Promise.reject('Pic get failed')
       return img
     })
-    return requestObj
   },
 }